fix(billing): skip batchPut if no space diff (#4)
Issue: https://github.com/storacha/project-tracking/issues/185

Potential Fix: if there are no records for that `.list({consumer:
delta.resource})` in the `consumerStore`, it will return an empty array,
then it won't generate any space diff to be stored, and it will throw
the error when we execute the `batchPut` call. The fix skips the
`batchPut` call when no diffs are available.